# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Waste Recycling Reward System API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: users
    description: User operations
  - name: waste
    description: Waste submission operations
  - name: admin
    description: Admin operations
  - name: stats
    description: Statistics operations
  - name: leaderboard
    description: Leaderboard operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register or update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterUserRequest"
      responses:
        "200":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logoutUser
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /users/profile:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get user profile with token balance
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/wallet:
    put:
      operationId: updateWalletAddress
      tags: [users]
      summary: Update user wallet address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWalletRequest"
      responses:
        "200":
          description: Wallet updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /waste/submit:
    post:
      operationId: submitWaste
      tags: [waste]
      summary: Submit recyclable waste
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitWasteRequest"
      responses:
        "201":
          description: Waste submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WasteSubmission"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /waste/submissions:
    get:
      operationId: getUserSubmissions
      tags: [waste]
      summary: Get user's waste submissions
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Submissions list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionsList"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/submissions:
    get:
      operationId: adminGetSubmissions
      tags: [admin]
      summary: Admin get all submissions
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: All submissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSubmissionsList"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/submissions/{id}/approve:
    post:
      operationId: approveSubmission
      tags: [admin]
      summary: Approve a waste submission and issue tokens
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submission approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WasteSubmission"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/submissions/{id}/reject:
    post:
      operationId: rejectSubmission
      tags: [admin]
      summary: Reject a waste submission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submission rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WasteSubmission"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/users:
    get:
      operationId: adminGetUsers
      tags: [admin]
      summary: Admin get all users
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"

  /admin/stats:
    get:
      operationId: adminGetStats
      tags: [admin]
      summary: Admin get system statistics
      responses:
        "200":
          description: System stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

  /stats/overview:
    get:
      operationId: getStatsOverview
      tags: [stats]
      summary: Get public statistics overview
      responses:
        "200":
          description: Statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsOverview"

  /stats/waste-types:
    get:
      operationId: getWasteTypeStats
      tags: [stats]
      summary: Get waste breakdown by type
      responses:
        "200":
          description: Waste type statistics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WasteTypeStat"

  /stats/monthly:
    get:
      operationId: getMonthlyStats
      tags: [stats]
      summary: Get monthly recycling statistics
      responses:
        "200":
          description: Monthly stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyStat"

  /leaderboard:
    get:
      operationId: getLeaderboard
      tags: [leaderboard]
      summary: Get top recyclers leaderboard
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        walletAddress:
          type: string
          nullable: true
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - createdAt

    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        walletAddress:
          type: string
          nullable: true
        role:
          type: string
          enum: [user, admin]
        tokenBalance:
          type: number
        totalWasteKg:
          type: number
        totalSubmissions:
          type: integer
        approvedSubmissions:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - tokenBalance
        - totalWasteKg
        - totalSubmissions
        - approvedSubmissions
        - createdAt

    RegisterUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        walletAddress:
          type: string
          nullable: true
      required:
        - name
        - email
        - password

    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password

    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required:
        - user
        - message

    UpdateWalletRequest:
      type: object
      properties:
        walletAddress:
          type: string
      required:
        - walletAddress

    WasteSubmission:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        wasteType:
          type: string
          enum: [Plastic, Paper, Metal, Glass, E-waste]
        weightKg:
          type: number
        recyclingCenterId:
          type: string
        photoUrl:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, approved, rejected]
        tokensAwarded:
          type: number
          nullable: true
        transactionHash:
          type: string
          nullable: true
        blockchainTimestamp:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - wasteType
        - weightKg
        - recyclingCenterId
        - status
        - createdAt
        - updatedAt

    SubmitWasteRequest:
      type: object
      properties:
        wasteType:
          type: string
          enum: [Plastic, Paper, Metal, Glass, E-waste]
        weightKg:
          type: number
          minimum: 0.1
        recyclingCenterId:
          type: string
        photoUrl:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - wasteType
        - weightKg
        - recyclingCenterId

    SubmissionsList:
      type: object
      properties:
        submissions:
          type: array
          items:
            $ref: "#/components/schemas/WasteSubmission"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - submissions
        - total
        - page
        - limit

    AdminSubmissionsList:
      type: object
      properties:
        submissions:
          type: array
          items:
            $ref: "#/components/schemas/WasteSubmission"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - submissions
        - total
        - page
        - limit

    AdminStats:
      type: object
      properties:
        totalUsers:
          type: integer
        totalSubmissions:
          type: integer
        pendingSubmissions:
          type: integer
        approvedSubmissions:
          type: integer
        rejectedSubmissions:
          type: integer
        totalWasteKg:
          type: number
        totalTokensIssued:
          type: number
        wasteByType:
          type: array
          items:
            $ref: "#/components/schemas/WasteTypeStat"
      required:
        - totalUsers
        - totalSubmissions
        - pendingSubmissions
        - approvedSubmissions
        - rejectedSubmissions
        - totalWasteKg
        - totalTokensIssued
        - wasteByType

    StatsOverview:
      type: object
      properties:
        totalWasteKg:
          type: number
        totalTokensIssued:
          type: number
        totalUsers:
          type: integer
        totalSubmissions:
          type: integer
      required:
        - totalWasteKg
        - totalTokensIssued
        - totalUsers
        - totalSubmissions

    WasteTypeStat:
      type: object
      properties:
        wasteType:
          type: string
        totalKg:
          type: number
        count:
          type: integer
      required:
        - wasteType
        - totalKg
        - count

    MonthlyStat:
      type: object
      properties:
        month:
          type: string
        totalKg:
          type: number
        totalTokens:
          type: number
        submissions:
          type: integer
      required:
        - month
        - totalKg
        - totalTokens
        - submissions

    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        userId:
          type: integer
        name:
          type: string
        tokenBalance:
          type: number
        totalWasteKg:
          type: number
        totalSubmissions:
          type: integer
      required:
        - rank
        - userId
        - name
        - tokenBalance
        - totalWasteKg
        - totalSubmissions