# lib/api-zod/src/generated/types/submissionsList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Waste Recycling Reward System API
 * OpenAPI spec version: 0.1.0
 */
import type { WasteSubmission } from "./wasteSubmission";

export interface SubmissionsList {
  submissions: WasteSubmission[];
  total: number;
  page: number;
  limit: number;
}