# lib/api-zod/src/generated/types/updateWalletRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Waste Recycling Reward System API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateWalletRequest {
  walletAddress: string;
}