# lib/api-zod/src/generated/types/wasteSubmission.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Waste Recycling Reward System API
 * OpenAPI spec version: 0.1.0
 */
import type { WasteSubmissionStatus } from "./wasteSubmissionStatus";
import type { WasteSubmissionWasteType } from "./wasteSubmissionWasteType";

export interface WasteSubmission {
  id: number;
  userId: number;
  userName?: string;
  wasteType: WasteSubmissionWasteType;
  weightKg: number;
  recyclingCenterId: string;
  photoUrl?: string | null;
  status: WasteSubmissionStatus;
  tokensAwarded?: number | null;
  transactionHash?: string | null;
  blockchainTimestamp?: string | null;
  notes?: string | null;
  createdAt: Date;
  updatedAt: Date;
}