# lib/db/src/schema/users.ts
import { pgTable, serial, text, timestamp, integer, real, pgEnum } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const roleEnum = pgEnum("role", ["user", "admin"]);

export const usersTable = pgTable("users", {
  id: serial("id").primaryKey(),
  name: text("name").notNull(),
  email: text("email").notNull().unique(),
  passwordHash: text("password_hash").notNull(),
  walletAddress: text("wallet_address"),
  role: roleEnum("role").notNull().default("user"),
  tokenBalance: real("token_balance").notNull().default(0),
  totalWasteKg: real("total_waste_kg").notNull().default(0),
  createdAt: timestamp("created_at").notNull().defaultNow(),
  updatedAt: timestamp("updated_at").notNull().defaultNow(),
});

export const insertUserSchema = createInsertSchema(usersTable).omit({
  id: true,
  tokenBalance: true,
  totalWasteKg: true,
  createdAt: true,
  updatedAt: true,
});

export const registerUserSchema = z.object({
  name: z.string().min(1),
  email: z.email(),
  password: z.string().min(6),
  walletAddress: z.string().optional().nullable(),
});

export const loginSchema = z.object({
  email: z.email(),
  password: z.string().min(1),
});

export type InsertUser = z.infer<typeof insertUserSchema>;
export type User = typeof usersTable.$inferSelect;